Convert SubcompileNode into anonymous class
